rename train_config to training_config
