+++ .github/workflows/publish-testpypi.yml
@@ -58,7 +58,7 @@ jobs:
           python - <<'PY'
           import importlib.metadata as metadata
 
-          import physicalai.train_config
+          import physicalai.training_config
           import physicalai.data
           import physicalai.policies
           import physicalai.train
