+++ library/src/physicalai/inference/adapters/pytorch.py
@@ -17,7 +17,7 @@ from physicalai.inference.adapters.base import RuntimeAdapter
 from physicalai.inference.adapters.registry import adapter_registry
 from physicalai.inference.manifest import Manifest
 
-from physicalai.train_config import import_class
+from physicalai.training_config import import_class
 from physicalai.data.observation import Observation
 from physicalai.export.backends import TorchExportParameters
