+++ library/src/physicalai/policies/act/config.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
-from physicalai.train_config import Config
+from physicalai.training_config import Config
 from physicalai.data import Feature  # noqa: TC001 - Needed at runtime for type hint resolution
