+++ library/src/physicalai/policies/groot/components/action_head.py
@@ -17,7 +17,7 @@ import torch.nn.functional as F  # noqa: N812
 from torch import nn
 from torch.distributions import Beta
 
-from physicalai.train_config.mixin import FromConfig
+from physicalai.training_config.mixin import FromConfig
 
 from .nn import CategorySpecificMLP, MultiEmbodimentActionEncoder
 from .transformer import get_dit_class, get_self_attention_transformer_class
