+++ library/src/physicalai/policies/pi0/policy.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING, Any, Literal, cast
 
 import torch
 
-from physicalai.train_config.mixin import FromConfig
+from physicalai.training_config.mixin import FromConfig
 from physicalai.export.mixin_policy import ExportablePolicyMixin
 from physicalai.policies.base import Policy
 from physicalai.train.utils import reformat_dataset_to_match_policy
