+++ library/src/physicalai/policies/smolvla/config.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from physicalai.train_config import Config
+from physicalai.training_config import Config
 
 
 @dataclass(frozen=True)
