+++ library/src/physicalai/train/trainer.py
@@ -14,7 +14,7 @@ import torch
 from lightning.pytorch.callbacks import BatchSizeFinder
 from lightning.pytorch.strategies import DDPStrategy
 
-from physicalai.train_config.instantiate import instantiate_obj_from_dict
+from physicalai.training_config.instantiate import instantiate_obj_from_dict
 from physicalai.train.callbacks import PolicyDatasetInteraction
