+++ library/src/physicalai/training_config/__init__.py
@@ -3,8 +3,8 @@
 
 """Configuration utilities for physicalai."""
 
-from physicalai.train_config.base import Config
-from physicalai.train_config.instantiate import import_class, instantiate_obj
-from physicalai.train_config.mixin import FromConfig, from_config
+from physicalai.training_config.base import Config
+from physicalai.training_config.instantiate import import_class, instantiate_obj
+from physicalai.training_config.mixin import FromConfig, from_config
 
 __all__ = ["Config", "FromConfig", "from_config", "import_class", "instantiate_obj"]
