+++ library/tests/unit/config/test_config.py
@@ -13,9 +13,9 @@ import numpy as np
 import pytest
 from pydantic import BaseModel
 
-from physicalai.train_config import Config, from_config
-from physicalai.train_config.instantiate import import_class, instantiate_obj
-from physicalai.train_config.mixin import FromConfig
+from physicalai.training_config import Config, from_config
+from physicalai.training_config.instantiate import import_class, instantiate_obj
+from physicalai.training_config.mixin import FromConfig
 
 # =============================================================================
 # Test Fixtures
