+++ library/tests/unit/policies/test_groot.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 
 import pytest
 import torch
-from physicalai.train_config import Config
+from physicalai.training_config import Config
 from physicalai.policies.groot import Groot, GrootConfig
 
 # ============================================================================ #
